requirements.txt
requests
fastapi
uvicorn
python-multipart
pyjwt
passlib
passlib[bcrypt]
bcrypt
pyyaml==3.13
Jinja2